+++ changelog.md
@@ -1,5 +1,6 @@
 ## v0.1.10
 * Added [aka export] to export alias groups into a project directory
+* Added [aka export --name ] to generate specific aka file
 * Added [aka proj] to show all the aliases for current project directory
 * Added [aka groups] to list all groups
 * Edited test_aka.rb